+++ packages/webapp/src/inductions/components/induction-profile-form.tsx
@@ -10,6 +10,8 @@ import {
     handleFileChange,
     Text,
 } from "_app";
+import { edenContractAccount, validUploadActions } from "config";
+
 import { NewMemberProfile } from "../interfaces";
 
 interface Props {
@@ -101,7 +103,9 @@ export const InductionProfileForm = ({
                         handleFileChange(
                             e,
                             "image",
-                            1_000_000,
+                            validUploadActions[edenContractAccount][
+                                "inductprofil"
+                            ].maxSize,
                             setUploadedImage
                         )
                     }
